server/requirements.txt
Flask
flask-cors
supabase
python-dotenv
gunicorn
transformers
torch
torchvision
torchaudio
accelerate
requests>=2.31.0
google-genai